// src/generated/api/inventory/models/listStockTransfersStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */

export type ListStockTransfersStatus =
  (typeof ListStockTransfersStatus)[keyof typeof ListStockTransfersStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ListStockTransfersStatus = {
  DRAFT: 'DRAFT',
  SUBMITTED: 'SUBMITTED',
  SHIPPED: 'SHIPPED',
  RECEIVED: 'RECEIVED',
} as const;